Improved Mint/Burn btoken interface
(1) Implements Mint/Burn btoken methods using immutable reference to LendingMarket in order to reduce costs

(2) Implements flash mint/repay btokens - for efficiency gains + better event handling

Flash mint/repay is praxeologically equivalent to the regular mint_btoken flow, with the benefit that allows it to return some bToken change. Given that deposits and swaps are subject to slippage, this logic has the benefit that it allows the user to mint BTokens and return the leftover BToken amounts in the repay phase. The alternative would require the user to burn back the change, which is more computationally expensive (leads to duplicated computation of compound interest, etc.). This also has the extra benefit of better event handling since we woun't be emitting false positive burn events.

Note: For extra safety we should consider making this flash logic `package` only and move the swap scripts inside the steamm package.